# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Music Platform API
  version: 1.0.0
  description: "OpenAPI 3 specification for the Music Platform backend. Base path assumed to be /api — if your app mounts the router on /api, prepend /api to paths."
servers:
  - url: https://localhost:3000
    description: Local development
paths:
...
  /media:
    post:
      summary: Upload media (multipart)
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                media_type:
                  type: string
    get:
      summary: List media
      parameters:
        - in: query
          name: owner_type
          schema: { type: string }
        - in: query
          name: owner_id
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: per_page
          schema: { type: integer }
  /media/{mediaId}:
    get:
      summary: Get media metadata
      parameters:
        - in: path
          name: mediaId
          required: true
          schema: { type: integer }
    patch:
      summary: Update media metadata (owner only)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                media_type: { type: string }
    post:
      summary: Replace media file (owner only)
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
    delete:
      summary: Delete media (owner only)
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []